Move compactEntityJson.js to root/static/scripts/edit/utility/
It's not used outside of the browser, so doesn't really make sense to have
outside root/static/scripts/.
