+++ root/static/scripts/edit/utility/compactEntityJson.js
